share string interner between scraper and remote_write
This commit exposes the string interner globally with a shared pool and
modifies both the scraper and remote_write to use that shared pool.

Note that the string_interner_zero_reference_releases_total now has a
different namespace and subsystem from what it was initially set to in
the RW code.

Signed-off-by: Robert Fratto <robertfratto@gmail.com>